engine: an indexed axis has no dim-element spelling
The ceteris-paribus wrap qualified a resolved subscript element as `dim·elem`
whatever kind of dimension the axis was. DimensionsContext::lookup resolves that
spelling only for a NAMED dimension, so an index into an indexed axis came out as
`d·1`, which resolves to nothing: the PREVIOUS-capture helper holding it failed to
compile and both of that target's link scores read a constant 0 behind Assembly
warnings. The qualification now goes through qualify_axis_element, the one
qualifier, which leaves an indexed axis's position verbatim -- and verbatim is
what resolves.

The resolution itself was right and is unchanged, which is worth recording
because it is easy to read the other way. A quoted numeric-named variable used as
an index -- `q["1"]` -- is not a runtime read of that variable:
normalize_subscripts3 declines it, but declining routes the subscript to the
dynamic lowering, whose first move is Dimension::get_offset, and that parses an
indexed axis's identifier text into a position. Measured on a running model, with
the variable holding two different values and the result unchanged. Refusing to
resolve it instead would make the wrap freeze the index, and a frozen index is a
dynamic one -- the partial would read whatever the variable held last step where
the equation reads position 1, turning a loud zero into a silent wrong row.

DimName.N remains a deliberate gap in the shared resolver, now pinned by a test
asserting that shape does not compile at all, with instructions for the day it
does.

ltm_augment.rs sat exactly on the per-file line cap, so the wrap's
subscript-index pass moves to a #[path]-mounted sibling -- a cohesive unit and a
pure move.
